requirements.txt
fastapi
uvicorn
spotipy
python-dotenv
pandas
numpy
librosa
scikit-learn
pytest
requests
